+++ ValidnatorXSD/Models/ResponseErrorsModel.cs
@@ -3,7 +3,7 @@
     public class ResponseErrorsModel
     {
         public string Message { get; set; }
-        public long? ColumnPos { get; set; }
-        public long RowPos { get; set; }
+        public int? ColumnPos { get; set; }
+        public int RowPos { get; set; }
     }
 }
